Refactor initialization process and add genome configuration
This commit refactors the initialization process in the `cli.py` file to include a new `init` command. The `init` command initializes the seqnado pipeline and sets up the correct environmental variables. It also adds the functionality to initialize genome files using a template file. The `get_genomes` function in the `helpers.py` file is modified to read the genome configuration from a JSON file and return the genome values. The commit also includes the addition of a new `init.sh` script that is used to initialize the environment for the SeqNado project. Lastly, a template file `genomes_template.json` is added to the `workflow/config` directory.